Hweave {highlight}	R Documentation

Weaving and Tangling with syntax highlighting

Description

Hweave and Htangle are similar to Sweave and Stangle, but they take advantage of the custom driver shipped with this package

Usage

Hweave(file, driver = HighlightWeaveLatex(), syntax = HweaveSyntaxNoweb,
  encoding = "", ...)

Htangle(file, driver = HighlightTangle(), syntax = HweaveSyntaxNoweb,
  encoding = "", ...)

Arguments

file
Path to Sweave source file

driver
The actual workhorse, see the Details section in Sweave

syntax
NULL or an object of class SweaveSyntax or a character string with its name. See the section Syntax Definition in Sweave

encoding
The default encoding to assume for file

...
Further arguments passed to the driver's setup function.

Details

These functions exist for the purpose of the \VignetteEngine option in vignette introduced in R 3.0.0

highlight loads the highlight vignette engine at load time. Client packages must declare to use it with the VignetteBuilder field in their DESCRIPTION file

The vignette engine looks for files matching the pattern "[.][hHrRsS]nw$" although in order to distinguish vignettes using this engine and the default Sweave engine, the recommandation is to use vignette with the ".Hnw" extension.

[Package highlight version 0.4.7.2 Index]