boxes_latex {highlight}	R Documentation

Creates the set of latex boxes

Description

This function returns the set of latex boxes definitions that should be included in the document preamble. The latex renderer includes these definitions automatically when the document argument is TRUE, but not otherwise.

Usage

boxes_latex()

Value

A character vector containing latex definitions for boxes used by the latex renderer

See Also

translator_latex translates text into markup that makes use of these boxes

[Package highlight version 0.4.7.2 Index]