getStyleFile {highlight}	R Documentation

helper function to get a style file

Description

helper function to get a style file

Usage

getStyleFile(name = "default", extension = "css")

Arguments

name
the name of the style file to look for

extension
the file extension (css, sty, or xterm)

Details

the search goes as follows: first the current working directory then the directory ~/.R/highlight, then the stylesheet directory in the installed package

Value

the name of the first file that is found, or NULL

[Package highlight version 0.4.7.2 Index]