header_html {highlight}	R Documentation

html renderer header and footer

Description

these functions build the header function and the footer function used by the html renderer

Usage

header_html(document, stylesheet)

footer_html(document)

Arguments

document
logical. If TRUE the built header and footer functions will return the beginning and end of a full html document. If FALSE, the built functions will only return the opening and closing <pre> tags.

stylesheet
stylesheet to use. See getStyleFile for details on where the stylesheet can be located.

Value

header and footer functions.

See Also

renderer_html uses these functions to create a renderer suitable for the renderer argument of highlight

Examples

h <- header_html( document = FALSE )
h()
h <- header_html( document = TRUE, stylesheet = "default")
h()
f <- footer_html( document = TRUE )
f()
f <- footer_html( document = FALSE )
f()

[Package highlight version 0.4.7.2 Index]