header_latex {highlight}	R Documentation

latex header and footer

Description

These functions return appropriate header and footer functions for the latex renderer

Usage

header_latex(document, styles, boxes, minipage = FALSE)

footer_latex(document, minipage = FALSE)

Arguments

document
logical. If TRUE the header and footer functions will create the full document (including preamble with boxes and styles)

styles
a vector of style definitions to include in the preamble if document is TRUE

boxes
a vector of boxes definitions to include in the preamble if document is TRUE

minipage
if TRUE, the highlighted latex is included in a minipage environment

Value

A function is returned, suitable for the header or footer argument of the latex renderer

Examples

h <- header_latex( document = FALSE )
h()
f <- footer_latex( document = FALSE )
f()

[Package highlight version 0.4.7.2 Index]