simple_detective {highlight}	R Documentation

Simple detective

Description

This detective only uses semantic information to make its investigation.

Usage

simple_detective(x, ...)

Arguments

x
output of the parser. The detective is only interested in the token column of the data.

...
ignored

Value

a vector of styles grouping similar tokens together

Examples

## Not run: 
p <- parse( text = deparse( jitter ), keep.source=TRUE )
simple_detective( p )

## End(Not run)

[Package highlight version 0.4.7.2 Index]