styler_assistant_latex {highlight}	R Documentation

latex styler assistant

Description

This function takes the output of the css.parser and produces latex style definitions from it.

Usage

styler_assistant_latex(x)

Arguments

x
output from css.parser

Details

The function create a new latex command for each css declaration, i.e. each item of the list x it is passed.

The assistant currently honours the following css settings: color, text-decoration:underline, font-weight:bold[er] and font-style:italic

Value

a vector of latex style definitions corresponding to (a subset of) the output of the parser

See Also

styler

[Package highlight version 0.4.7.2 Index]