html renderer header and footer

these functions build the header function and the footer function used by the html renderer

header_html(document, stylesheet)

footer_html(document)

Arguments

document

logical. If TRUE the built header and footer functions will return the beginning and end of a full html document. If FALSE, the built functions will only return the opening and closing <pre> tags.

stylesheet

stylesheet to use. See getStyleFile for details on where the stylesheet can be located.

Value

header and footer functions.

See also

Examples

h <- header_html( document = FALSE )
h()
#> [1] "<pre>\n"
h <- header_html( document = TRUE, stylesheet = "default")
#> Warning: one argument not used by format '%s$'
h()
#>  [1] "<html>\n<head>\n<style type=\"text/css\">\n"
#>  [2] ".number{\n"                                 
#>  [3] "\tcolor: rgb(21,20,181) ;\n"                
#>  [4] "}\n"                                        
#>  [5] "\n"                                         
#>  [6] ".functioncall{\n"                           
#>  [7] "\tcolor: red ;\n"                           
#>  [8] "}\n"                                        
#>  [9] "\n"                                         
#> [10] ".string{\n"                                 
#> [11] "\tcolor: rgb(153,153,255) ;\n"              
#> [12] "}\n"                                        
#> [13] "\n"                                         
#> [14] ".keyword{\n"                                
#> [15] "\tcolor: black;\n"                          
#> [16] "}\n"                                        
#> [17] "\n"                                         
#> [18] ".argument{\n"                               
#> [19] "\tcolor: rgb( 177,63,5) ;\n"                
#> [20] "}\n"                                        
#> [21] "\n"                                         
#> [22] ".comment{\n"                                
#> [23] "\tcolor: rgb( 204,204,204) ;\n"             
#> [24] "}\n"                                        
#> [25] "\n"                                         
#> [26] ".roxygencomment{\n"                         
#> [27] "\tcolor: rgb(0,151,255);\n"                 
#> [28] "}\n"                                        
#> [29] "\n"                                         
#> [30] ".formalargs{\n"                             
#> [31] "\tcolor: rgb(18,182,18);\n"                 
#> [32] "}\n"                                        
#> [33] "\n"                                         
#> [34] ".eqformalargs{\n"                           
#> [35] "\tcolor: rgb(18,182,18);\n"                 
#> [36] "}\n"                                        
#> [37] "\n"                                         
#> [38] ".assignement{\n"                            
#> [39] "\tcolor: rgb(55,55,98);\n"                  
#> [40] "}\n"                                        
#> [41] "\n"                                         
#> [42] ".package{\n"                                
#> [43] "\tcolor: rgb(150,182,37);\n"                
#> [44] "}\n"                                        
#> [45] "\n"                                         
#> [46] ".slot{\n"                                   
#> [47] "\tfont-style:italic;\n"                     
#> [48] "}\n"                                        
#> [49] "\n"                                         
#> [50] ".symbol{\n"                                 
#> [51] "\tcolor: black ;\n"                         
#> [52] "}\n"                                        
#> [53] "\n"                                         
#> [54] ".prompt{\n"                                 
#> [55] "\tcolor: black ;\n"                         
#> [56] "}\n"                                        
#> [57] "\n"                                         
#> [58] ".line{\n"                                   
#> [59] "    color: gray ;   \n"                     
#> [60] "}\n"                                        
#> [61] "</style>\n</head>\n<body>\n<pre>\n"         
f <- footer_html( document = TRUE )
f()
#> [1] "\n</pre>\n</body>\n</html>\n"
f <- footer_html( document = FALSE )
f()
#> [1] "\n</pre>\n"